+++ packages/swirl-components/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @getflip/swirl-components
 
+## 0.53.0
+
+### Minor Changes
+
+- [`c102537`](https://github.com/getflip/swirl/commit/c1025377f595c8a3fd9a9686858cdac6d9ff4e87)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add "font-family" prop to
+  swirl-text
+
+## 0.52.0
+
+### Minor Changes
+
+- [`4d0fe1f`](https://github.com/getflip/swirl/commit/4d0fe1f1b8882f72679ddb482f24fb0b3b659571)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add content slot to swirl-toast;
+  deprecate "content" prop
+
+## 0.51.0
+
+### Minor Changes
+
+- [`c11c3f4`](https://github.com/getflip/swirl/commit/c11c3f4e789c1b310d1006508b77cde2701babc3)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add input event to swirl-search
+
+* [`d5d518a`](https://github.com/getflip/swirl/commit/d5d518aa421b9230c6f36619db24cddb5ef38271)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Replace swirl-pagination select
+  with input field
+
+- [`02614e9`](https://github.com/getflip/swirl/commit/02614e919b97a799a8e81633e4b3d89e4e8b6605)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Allow HTML inside swirl-toast
+  content
+
 ## 0.50.2
 
 ### Patch Changes